+++ CHANGELOG.md
@@ -7,6 +7,20 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Added
+
+- Add methods to modify the text node of an element:
+
+  - `Elements.textNode(elemental2.dom.Element, java.lang.String)`
+  - `Elements.textNode(org.jboss.elemento.IsElement<E>, java.lang.String)`
+
+- Add overloaded methods to bind events in `HasElement`:
+
+  - `HasElement.on(EventType<V,?>, boolean, EventCallbackFn<V>)`
+  - `HasElement.on(EventType<V,?>, AddEventListenerOptions, EventCallbackFn<V>)`
+
+- Add `Logger.isEnabled(Level)` to check if a level is enabled for a logger.
+
 ## [1.5.0] - 2024-05-06
 
 ### Added